add app properties & dataanalysisservice
