--- library/src/MyService.java
@@ -1,19 +1,0 @@
-package com.example.multimodule.service;
-
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.stereotype.Service;
-
-@Service
-@EnableConfigurationProperties(ServiceProperties.class)
-public class MyService {
-
-  private final ServiceProperties serviceProperties;
-
-  public MyService(ServiceProperties serviceProperties) {
-    this.serviceProperties = serviceProperties;
-  }
-
-  public String message() {
-    return this.serviceProperties.getMessage();
-  }
-}
